+++ docs/changelog.md
@@ -5,6 +5,8 @@
     * Masonry Brick Slab [(#194)](https://github.com/codetaylor/pyrotech-1.12/issues/194)
     * Refractory Brick Stairs
     * Refractory Brick Slab
+    * Valid refractory blocks:
+        * Double Refractory Brick Slab
     * Ore Compat:
         * Bloom yield and slag bloom yield can now be adjusted in the ore compat json files [(#310)](https://github.com/codetaylor/pyrotech-1.12/issues/310)
     * Config option to: